refactor: extract types in separate file
